requirements.txt
selene==2.0.0rc4
pytest~=7.4.3
allure-pytest
python-dotenv
pydantic-settings
